PFE-backend/requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-ipware==7.0.1
django-recaptcha==4.1.0
django-sslserver==0.22
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-recaptcha==4.0.2
drf-yasg==1.21.10
fonttools==4.57.0
fpdf2==2.8.3
inflection==0.5.1
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
python-ipware==3.0.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
whitenoise==6.9.0
